Introduce RegistryValue & DefaultedTaggable, move some things around
